Start a dbus daemon if there isn't already one present
Usually, we rely on the X server to give us the path to the dbus server.
Since this program needs to be run on machines headlessly, we can't rely
on that. This usually means that nothing starts dbus, and we have to
start our own dbus daemon, and then close it down when we're done with
it.